src/requirements.txt
asgiref==3.6.0
Django==3.2
django-language-field==0.0.3
django-multiplechoicefield==1.0.1
django-multiselectfield==0.1.12
djangorestframework==3.14.0
pytz==2022.7.1
sqlparse==0.4.3
